+++ cotests/bench/group/utils.py
@@ -1,0 +1,19 @@
+import asyncio
+import inspect
+
+from ..typ import RunResult
+
+
+def try_to_run(t) -> 'RunResult':
+    if t and inspect.iscoroutine(t):
+        # try to run
+        try:
+            asyncio.get_running_loop()
+        except RuntimeError:
+            # print('Run in new loop')
+            asyncio.run(t)
+        else:
+            # print('Cannot run. Return coroutine')
+            return t
+    # else:
+    #     print('No coroutines')
